database-from-scratch/src/test/java/com/mavcorp/database/dao/impl/BookDaoImplTests.java
package com.mavcorp.database.dao.impl;

import com.mavcorp.database.TestDataUtil;
import com.mavcorp.database.domain.Book;
import org.junit.jupiter.api.Test;
import org.junit.jupiter.api.extension.ExtendWith;
import org.mockito.ArgumentMatchers;
import org.mockito.InjectMocks;
import org.mockito.Mock;
import org.mockito.junit.jupiter.MockitoExtension;
import org.springframework.jdbc.core.JdbcTemplate;

import static org.mockito.ArgumentMatchers.eq;
import static org.mockito.Mockito.verify;

@ExtendWith(MockitoExtension.class)
public class BookDaoImplTests {
    @Mock
    private JdbcTemplate jdbcTemplate;

    @InjectMocks
    private BookDaoImpl underTest;

    @Test
    public void testThatCreateBookGeneratesCorrectSql(){
        Book book = TestDataUtil.createTestBookA();
        underTest.create(book);

        verify(jdbcTemplate).update(
                eq("INSERT INTO books (isbn,title,author_id) VALUES (?,?,?)"),
                eq("831-231-31-3"),eq("Book1"),eq(1L)
        );
    }

    @Test
    public void testThatFindOneBookGeneratesCorrectSql(){
        underTest.findOne("831-231-31-3");
        verify(jdbcTemplate).query(
                eq("SELECT * FROM books WHERE isbn = ? LIMIT 1"),
                ArgumentMatchers.<BookDaoImpl.BookRowMapper>any(),
                eq("831-231-31-3")
        );
    }

    @Test
    public void testThatFindManyBooksGeneratesCorrectSql(){
        underTest.findMany();
        verify(jdbcTemplate).query(
                eq("SELECT * FROM books"),
                ArgumentMatchers.<BookDaoImpl.BookRowMapper>any()
        );
    }

    @Test
    public void testThatUpdateBookGeneratesCorrectSql(){
        Book book = TestDataUtil.createTestBookA();
        underTest.update("831-231-31-3",book);

        verify(jdbcTemplate).update(
                eq("UPDATE books SET isbn = ?, title = ?, author_id = ? WHERE isbn = ?"),
                eq("831-231-31-3"),eq("Book1"),eq(1L),eq("831-231-31-3")
        );
    }

    @Test
    public void testThatDeleteBookGeneratesCorrectSql(){
        underTest.delete("831-231-31-3");
        verify(jdbcTemplate).update(
                eq("DELETE FROM books WHERE isbn = ?"),
                eq("831-231-31-3")
        );
    }

}

